Added a basic internal profiler and UI view. Fixed high precision issues with adjoins (which now work correctly at high res). Added various Classic Renderer counters (for the profiler view). Added profiler "zones" to the classic renderer to track performance better. Fixed setCamera(), (x,y,z) should be fixed16 type. Moved some sector update setup to sector_copy() so it only happens once.

+++ TheForceEngine/TFE_Game/level.cpp
@@ -136,7 +136,7 @@ namespace TFE_Level
 				if (oclass != CLASS_SPRITE && oclass != CLASS_FRAME && oclass != CLASS_3D && oclass != CLASS_SOUND) { continue; }
 				// Skip objects that are not in sectors.
 				if (sectorId < 0) { continue; }
-				if (sectorId >= sectorCount)
+				if (sectorId >= s32(sectorCount))
 				{
 					TFE_System::logWrite(LOG_ERROR, "Level", "Sector ID for object is invalid: %d.", sectorId);
 					continue;
